refactor(masonry): [brick] seperate text measurement utility
